+++ src/scene_server/admin_server/imports.go
@@ -50,5 +50,5 @@ import (
 	_ "configcenter/src/scene_server/admin_server/upgrader/y3.10.202105251041"
 	_ "configcenter/src/scene_server/admin_server/upgrader/y3.10.202105261459"
 	_ "configcenter/src/scene_server/admin_server/upgrader/y3.10.202106031151"
-	_ "configcenter/src/scene_server/admin_server/upgrader/y3.10.202108202015"
+	_ "configcenter/src/scene_server/admin_server/upgrader/y3.10.202109072015"
 )
